requirements.txt
# production
fastapi~=0.103.1
uvicorn~=0.23.2
requests~=2.31.0
python-dotenv~=1.0.0
python-multipart~=0.0.5
pydantic~=2.4.2
httpx~=0.24.1
starlette~=0.27.0
sqlalchemy
pydantic-settings
passlib


# tests
pytest~=7.4.2
pytest-cov~=4.1.0
PyHamcrest~=2.0.4
busypie~=0.5.1
pytest-asyncio~=0.21.0
mimicker~=2.0.3

# helpers
ipython~=8.16.1
flake8~=6.1.0
pre-commit~=3.4.0
mccabe~=0.7.0
jq
